web-service/requirements.txt
google-cloud-storage
Flask
gunicorn